Include lig in and sign up in one component, add style files
