+++ src/App.js
@@ -1,15 +1,13 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import PrivateRoute from './PrivateRoute';
-import Login from './Login';
-import SignUp from './SignUp';
+import Login from './Components/Login';
 
 function App() {
   return (
     <BrowserRouter>
       <Switch>
-        <Route exact path="/login" component={Login} />
-        <Route exact path="/users" component={SignUp} />
+        <Route exact path={['/users/login', '/users/sign-up']} component={Login} />
         <PrivateRoute />
       </Switch>
     </BrowserRouter>
