+++ src/logic/storage.js
@@ -1,10 +1,10 @@
 const saveToken = (token) => {
   localStorage.token = JSON.stringify(token);
-}
+};
 
 const loadToken = () => {
-  const token = JSON.parse(localStorage.token)
-  return token
-}
+  const token = JSON.parse(localStorage.token);
+  return token;
+};
 
-export { saveToken, loadToken }
+export { saveToken, loadToken };
